+++ code/game-server/server/GameSettings.ts
@@ -1,0 +1,3 @@
+/** Preferences and Settings for this lobby */
+
+export class GameSettings { }
