+++ code/game-server/server/Player.ts
@@ -1,0 +1,12 @@
+import { AccountId } from "./types";
+
+
+export class Player {
+  displayName: string;
+  linkedAcccountId?: AccountId;
+
+  constructor(displayName: string, linkedAcccountId: AccountId | undefined) {
+    this.displayName = displayName;
+    this.linkedAcccountId = linkedAcccountId;
+  }
+}
