+++ code/game-server/server/Room.ts
@@ -1,0 +1,23 @@
+import { RoomId, JoinCode } from "./types";
+import { Player } from "./Player";
+import { GameSettings } from "./GameSettings";
+
+
+export class Room {
+  readonly roomId: RoomId;
+  /**
+   * The code which players can join the room with.
+   *
+   * Generated from {@link roomId} using {@link Sqids}
+   */
+  readonly joinCode: JoinCode;
+
+  players: Map<string, Player> = new Map<string, Player>();
+  gameState: any = undefined;
+  settings: GameSettings = new GameSettings();
+
+  constructor(roomId: RoomId, joinCode: JoinCode) {
+    this.roomId = roomId;
+    this.joinCode = joinCode;
+  }
+}
