+++ code/game-server/server/types.ts
@@ -1,0 +1,7 @@
+// TODO export to all folders / make global across services
+
+export type ServerId = number;
+export type RoomId = number;
+export type JoinCode = string;
+
+export type AccountId = string;
